requirements.txt
torchvision==0.9.1+cpu
numpy==1.16.5
torch==1.8.1+cpu
matplotlib==3.1.1
gym==0.17.2
Pillow==8.3.1
